+++ deployments/chef/spec/spec_helper.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+require 'chefspec'
+require 'chefspec/berkshelf'
+
+require 'shared_examples'